backend/requirements-dev.txt
pytest-cov==5.0.0
freezegun==1.5.1
jsonschema==4.23.0
